WP: Update pattern matching for gathering results
Why3 1.x.x modified the result pattern by removing a space between file
information and the colon (french vs. english punctuation).
This change caused the regex to break.

The previous pattern in 0.xx had the version number in parenthesis,
but now the name and version are only separated by a space instead. If
the prover has support for counter-examples then it will stand in
after the version number in parenthesis.